+++ fairseat-backend/fairseat-marketplace/src/main/java/com/pfa/fairseatmarketplace/dto/ResaleListingResponseDTO.java
@@ -1,6 +1,7 @@
 package com.pfa.fairseatmarketplace.dto;
 
 import com.pfa.fairseatmarketplace.domain.ListingStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,7 +10,7 @@ public record ResaleListingResponseDTO(
         UUID ticketId,
         String sellerId,
         Long gameId,
-        Double askingPrice,
+        BigDecimal askingPrice,
         ListingStatus status,
         LocalDateTime listedAt
 ) {}
